V1.25.0版本更新

+++ easyink-wecom/src/main/java/com/easyink/wecom/domain/dto/form/BatchUpdateGroupDTO.java
@@ -32,7 +32,7 @@ public class BatchUpdateGroupDTO {
      * 表单id列表
      */
     @NotEmpty(message = "请选择要修改的表单")
-    private List<Integer> formIdList;
+    private List<Long> formIdList;
 
     /**
      * 校验
